Extract contracts_routes.py from orgs_routes.py
Contracts had grown to roughly 60% of `orgs_routes.py` — ~676 of 1,138 lines
and about half its ~51 endpoints — arriving as one contiguous 561-line block
plus the typeahead pair and three CrudSpecs. They are no longer org metadata
in the way institutions and areas of interest are: they have their own page,
their own detail card, their own reload event, and an award-source
integration behind them.

The tell that the module wanted splitting: `blueprint.py`'s /admin/contracts
page route reached into `orgs_routes` for a *private* helper through a
function-local import, added purely to dodge the tail-import cycle. The
helper is now public (`active_contract_sources`) and owned by the module the
page belongs to. The import stays function-local — `blueprint.py` imports the
route modules at its tail, so that is structural, not incidental.

NSF programs come along: every one of their surfaces is reached from a
contract. The Organizations card keeps its NSF Programs tab.

Follows the precedent of resources_routes.py / facilities_routes.py /
rate_limits_routes.py / configuration_routes.py, registered by the same
one-line tail import.

Pure code motion — no endpoint name, URL rule, method, permission or template
changes. `tests/unit/test_route_map_parity.py` passes **without regenerating
the snapshot**, which is the gate. `make perf` 22 passed, no query-count
change. 1,138 lines becomes 448 + 754.

Also renumbers a duplicated `# 8.` section header in the perf suite.
